+++ src/main/java/com/samsamhajo/deepground/qna/question/service/QuestionService.java
@@ -1,23 +1,33 @@
 package com.samsamhajo.deepground.qna.question.service;
 
+import com.samsamhajo.deepground.feed.feed.model.FetchFeedResponse;
+import com.samsamhajo.deepground.feed.feed.model.FetchFeedsResponse;
+import com.samsamhajo.deepground.feed.feedshared.model.FetchSharedFeedResponse;
 import com.samsamhajo.deepground.member.entity.Member;
 import com.samsamhajo.deepground.member.repository.MemberRepository;
+import com.samsamhajo.deepground.qna.answer.dto.AnswerCreateRequestDto;
+import com.samsamhajo.deepground.qna.answer.dto.AnswerCreateResponseDto;
+import com.samsamhajo.deepground.qna.answer.service.AnswerService;
 import com.samsamhajo.deepground.qna.question.Dto.*;
 import com.samsamhajo.deepground.qna.question.entity.Question;
+import com.samsamhajo.deepground.qna.question.entity.QuestionMedia;
 import com.samsamhajo.deepground.qna.question.entity.QuestionTag;
 import com.samsamhajo.deepground.qna.question.exception.QuestionErrorCode;
 import com.samsamhajo.deepground.qna.question.exception.QuestionException;
+import com.samsamhajo.deepground.qna.question.repository.QuestionMediaRepository;
 import com.samsamhajo.deepground.qna.question.repository.QuestionRepository;
 import com.samsamhajo.deepground.qna.question.repository.QuestionTagRepository;
 import com.samsamhajo.deepground.techStack.entity.TechStack;
 import com.samsamhajo.deepground.techStack.repository.TechStackRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
-import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -29,6 +39,9 @@ public class QuestionService{
     private final TechStackRepository techStackRepository;
     private final MemberRepository memberRepository;
     private final QuestionTagService questionTagService;
+    private final AnswerService answerService;
+    private final QuestionTagService tagService;
+    private final QuestionMediaRepository questionMediaRepository;
 
     //질문 생성
     @Transactional
@@ -68,10 +81,15 @@ public class QuestionService{
     @Transactional
     public Long deleteQuestion(Long questionId, Long memberId) {
 
-        //TODO : question을 작성한 멤버가 맞는지, 삭제권한 있는지 추후 로직 작성
+        Member member = memberRepository.findById(memberId).orElseThrow(
+                ()-> new QuestionException(QuestionErrorCode.QUESTION_MEMBER_MISMATCH));
 
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new QuestionException(QuestionErrorCode.QUESTION_NOT_FOUND));
+
+        if(!question.getMember().getId().equals(memberId)) {
+            throw new QuestionException(QuestionErrorCode.QUESTION_MEMBER_MISMATCH);
+        }
 
         questionTagRepository.deleteAllByQuestionId(questionId);
         questionMediaService.deleteQuestionMedia(questionId);
@@ -85,7 +103,7 @@ public class QuestionService{
     public QuestionUpdateResponseDto updateQuestion(QuestionUpdateRequestDto questionUpdateRequestDto, Long memberId) {
 
         Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("회원을 찾을 수 없습니다."));
+                .orElseThrow(() -> new QuestionException(QuestionErrorCode.QUESTION_MEMBER_MISMATCH));
 
         if(!StringUtils.hasText(questionUpdateRequestDto.getTitle())) {
             throw new QuestionException(QuestionErrorCode.QUESTION_TITLE_REQUIRED);
@@ -97,6 +115,10 @@ public class QuestionService{
 
         Question question = questionRepository.findById(questionUpdateRequestDto.getQuestionId())
                 .orElseThrow(()-> new QuestionException(QuestionErrorCode.QUESTION_NOT_FOUND));
+
+        if(!question.getMember().getId().equals(memberId)) {
+            throw new QuestionException(QuestionErrorCode.QUESTION_MEMBER_MISMATCH);
+        }
 
         questionTagRepository.deleteAllByQuestionId(question.getId());
 
@@ -118,8 +140,10 @@ public class QuestionService{
         );
 
         question.updateQuesiton(questionUpdateRequestDto.getTitle(), questionUpdateRequestDto.getContent());
-        questionMediaService.deleteQuestionMedia(question.getId());
         updateQuestionMedia(questionUpdateRequestDto, question);
+        questionMediaService.deleteQuestionMedia(question.getId());
+
+
 
         return questionUpdateResponseDto;
 
@@ -133,12 +157,12 @@ public class QuestionService{
     }
 
     @Transactional
-    public QuestionUpdateStatusResponseDto updateQuestionStatus(QuestionUpdateStatusRequestDto questionUpdateStatusRequestDto, Long memberId) {
+    public QuestionUpdateStatusResponseDto updateQuestionStatus(QuestionUpdateStatusRequestDto questionUpdateStatusRequestDto, Long memberId, Long questionId) {
 
-        Member member = memberRepository.findById(questionUpdateStatusRequestDto.getMemberId()).orElseThrow(
-                ()-> new IllegalArgumentException("존재하는 사용자가 아닙니다."));
+        Member member = memberRepository.findById(memberId).orElseThrow(
+                ()-> new QuestionException(QuestionErrorCode.QUESTION_MEMBER_MISMATCH));
 
-        Question question = questionRepository.findById(questionUpdateStatusRequestDto.getQuestionId()).orElseThrow(
+        Question question = questionRepository.findById(questionId).orElseThrow(
                 () -> new QuestionException(QuestionErrorCode.QUESTION_NOT_FOUND));
 
         if(!member.getId().equals(memberId)) {
@@ -152,5 +176,50 @@ public class QuestionService{
                 memberId
         );
     }
+
+
+    @Transactional(readOnly = true)
+    public QuestionListResponseDto getQuestions(Pageable pageable) {
+        Page<Question> questionPage = questionRepository.findAll(pageable);
+
+        List<QuestionSummaryDto> summaries = questionPage.stream()
+                .map(question -> {
+                    List<String> teckStacks = tagService.getStackNamesByQuestionId(question.getId());
+                    int answerCount = answerService.countAnswersByQuestionId(question.getId());
+
+                    return QuestionSummaryDto.of(question, teckStacks, answerCount);
+                }).toList();
+
+        return QuestionListResponseDto.of(summaries, questionPage.getTotalPages());
+    }
+
+    @Transactional(readOnly = true)
+    public QuestionDetailResponseDto getQuestionDetail(Long questionId, Long memberId) {
+        Question question = questionRepository.findById(questionId)
+                .orElseThrow(() -> new QuestionException(QuestionErrorCode.QUESTION_NOT_FOUND));
+        List<QuestionMedia> questionMedia = questionMediaRepository.findAllByQuestionId(questionId);
+        int answerCount = answerService.countAnswersByQuestionId(questionId);
+        List<String> techStacks = tagService.getStackNamesByQuestionId(questionId);
+        List<String> imageUrls = questionMedia.stream()
+                .map(QuestionMedia::getQuestionContentUrl)
+                .collect(Collectors.toList());
+
+        List<AnswerCreateResponseDto> answers = answerService.getAnswersByQuestionId(questionId);
+
+
+        return QuestionDetailResponseDto.of(
+                question.getId(),
+                question.getTitle(),
+                question.getContent(),
+                question.getMember().getId(),
+                question.getMember().getNickname(),
+                techStacks,
+                answerCount,
+                question.getQuestionStatus(),
+                imageUrls,
+                answers
+        );
+    }
+
 
 }
